==> Long-Run-Impacts-of-Recession/scripts/01-download_data.R <==
#### Preamble ####
# Purpose: Downloads and saves Data
# Author: Heyucheng Zhang, Jiwon Choi
# Date: 12 February 2024 
# Contact: heyucheng.zhang@mail.utoronto.ca, jwon.choi@mail.utoronto.ca
# License: MIT
# Pre-requisites: 00-simulate_data.R

#### Download data and save data ####

# We can't download the data using the R script since they are behind a sign-in.
# Replication URL: https://www.openicpsr.org/openicpsr/project/122001/version/V1/view
# Replication/data URL: https://www.openicpsr.org/openicpsr/project/122001/version/V1/view?path=/openicpsr/122001/fcr:versions/V1/replication/data&type=folder
# We downloaded three raw data files from the replication/data URL. They were "BLS_all_archived_counties7689aa (UNOFFICIAL).xlsx", "CA1_1969_2013_ALL.csv", "CA25_1969_2000_ALL.csv".
# We saved these three files in inputs/data/.
